优化部分代码实现
重构下载任务逻辑，优化日志与异常处理

重构代码以提升模块化与可维护性：
- 引入任务范围日志记录器，替代全局日志记录器。
- 将 `Result<T>` 替换为 `Result<T, ErrorCode>`，统一返回类型。
- 简化任务操作方法返回值，改为布尔值。
- 优化线程管理，将线程 ID 类型从 `int` 改为 `byte`。
- 增强异常处理，避免任务进入死循环并记录异常。
- 删除冗余代码，提升代码简洁性与可读性。

+++ MultithreadDownload/Threads/DownloadThread.cs
@@ -16,7 +16,7 @@ namespace MultithreadDownload.Threading
         /// <summary>
         /// The ID of the download thread. This is used to identify the thread in the download task.
         /// </summary>
-        public int ID { get; private set; }
+        public byte ID { get; private set; }
 
         /// <summary>
         /// The state of the download thread. This indicates whether the thread is running, paused, or stopped.
@@ -77,7 +77,7 @@ namespace MultithreadDownload.Threading
         /// </remarks>
         private readonly Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> _work;
 
-        public DownloadThread(int id, IDownloadContext downloadContext, string fileSegmentPath, Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> work)
+        public DownloadThread(byte id, IDownloadContext downloadContext, string fileSegmentPath, Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> work)
         {
             // Initialize the properties
             ID = id;
