+++ MultithreadDownload/Threads/DownloadThreadFactory.cs
@@ -9,7 +9,7 @@ namespace MultithreadDownload.Threading
 {
     public class DownloadThreadFactory : IDownloadThreadFactory
     {
-        public IDownloadThread Create(int id, IDownloadContext downloadContext, string fileSegmentPath, Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> work)
+        public IDownloadThread Create(byte id, IDownloadContext downloadContext, string fileSegmentPath, Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> work)
         {
             return new DownloadThread(id, downloadContext, fileSegmentPath, work);
         }
