+++ MultithreadDownload/Threads/IDownloadThreadFactory.cs
@@ -9,6 +9,6 @@ namespace MultithreadDownload.Threading
 {
     public interface IDownloadThreadFactory
     {
-        IDownloadThread Create(int id, IDownloadContext downloadContext,string fileSegmentPath, Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> work);
+        IDownloadThread Create(byte id, IDownloadContext downloadContext,string fileSegmentPath, Func<Stream, Stream, IDownloadThread, Result<bool, DownloadError>> work);
     }
 }
